hwc: Apply Default Mode only after Bootanimation is really completed
Property "service.bootanim.exit" was read before applying Default
Mode. This property is set in SurfaceFlinger when boot finish is
triggered which is happening before Bootanimation is completed
which results into some of the Bootanimation frames with default
mode settings.

Replace this property with "init.svc.bootanim" which is checked to
be "stopped" before applying default mode. This results into first
frame of homescreen to be with default mode settings.

CRs-Fixed: 805499
Change-Id: I23298a8c61b9a5bda239839d9f8ca9ecc3a6ab59